Update README.md to fix windows build instructions
In order for the node-pre-gyp rebuild function to work, the source files must also be located in the ./node_modules/.bin/ folder.  Hence, I have modified the instructions to include a step in which a user copies the required files into the ./node_modules/.bin/ so that the node-pre-gyp rebuild function can be successfully executed in the ./node_modules/.bin/ folder.